feat: remove unused LoginPreview component; streamline user authentication flow
