+++ frontend/src/pages/user/UserInputOTP.tsx
@@ -4,9 +4,7 @@ import OTPComponent from "@/components/OTPComponent"
 const UserInputOTP = () => {
     return (
         <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
-
             <div className="w-full max-w-sm">
-
                 <OTPComponent />
             </div>
         </div>
